requirements.txt
bullmq
python-dotenv
langchain-google-genai
langchain-core
